+++ src/main/java/com/zhumagulorken/cinema/showtime/service/MovieService.java
@@ -46,6 +46,9 @@ public class MovieService {
     }
 
     public void deleteMovie(Long id) {
+        if (!movieRepository.existsById(id)) {
+            throw new RuntimeException("Movie not found");
+        }
         movieRepository.deleteById(id);
     }
